requirements.txt
gradio>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
lightrag-hku>=0.0.1
numpy>=1.24.0
tqdm>=4.65.0
openai>=1.0.0
tiktoken>=0.5.0
